backend/requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.1.1
requests==2.32.3

